# docs/swagger.yaml
swagger: '2.0'
info:
  title: Somosport Core API
  description: Basic description of services provided by Somosport Core. Online editor & validator available at  http://editor.swagger.io/
  version: "1.0"
# the domain of the service
# host: ss-core-dev.herokuapp.com
#host: localhost:3000
# array of all schemes that your API supports
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /api/v1.0
produces:
  - application/json

##################################################
#
#  _____                 _
# /  ___|               (_)
# \ `--.  ___ _ ____   ___  ___ ___  ___
#  `--. \/ _ \ '__\ \ / / |/ __/ _ \/ __|
# /\__/ /  __/ |   \ V /| | (_|  __/\__ \
# \____/ \___|_|    \_/ |_|\___\___||___/
#
#
###################################################
paths:
...
  /user/team:
    get:
      summary: Returns a list of teams on which this user is an owner
      description: Returns a list of teams on which this user is an owner
      tags:
        - User
      parameters:
        - name: Authorization-Token
          in: header
          type: string
          required: true
      responses:
        200:
          description: User team list
        403:
          description: Invalid Authorization-Token or no Authorization-Token received
        404:
          description: No teams found
...
  /competition/team:
    get:
      summary: returns the list of participants in a competition
      description: returns the list of participants in a competition
      tags:
        - Competition
      produces:
      - application/json
      parameters:
      - name: page
        in: query
        type: integer
        description: Current page. Defaults to 1
      - name: page_size
        in: query
        type: integer
        description: Number of registers to return. Defaults to 10
      - name: sort
        in: query
        type: string
        description: Database name of the field to order by. Multiple fields allowed, separated by comma. Add a minus sign (-) to set order to desc. Defaults to 'asc'. Ex. 'created_at,-name'
      - name: competition_id
        in: query
        type: integer
        description: Filters competition by id
      - name: season_id
        in: query
        type: integer
        description: Filters competition's season by id. If found, returns only seasons with the given season_id
      - name: season_year
        in: query
        type: integer
        description: Filters seasons by init_at and end_at, from jan 1st <year> to dec 31th <year>
      - name: category_id
        in: query
        type: integer
        description: Filters seasons' categories by id. If found, returns only categories with the given category_id
      - name: team_status_id
        in: query
        type: integer
      responses:
        200:
          description: result
      security: []
...
  /category/{category_id}/team/{team_id}/player:
    get:
      summary: Gets all the players that participate on a team in a category. Also known as summoned players
      description: Gets all the players that participate on a team in a category. Also knownas summoned players
      operationId: get_category_team_player_
      produces:
      - application/json
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
        - name: team_id
          in: path
          required: true
          type: integer
      tags:
        - Player
        - Category
        - Team
      responses:
        200:
          description: pending
  /category/{category_id}/team/{team_id}/player/{player_id}:
    get:
      summary: REVISAR Gets the players that participate on a team in a category
      description: duplicado? Get players by category_id and team_id
      operationId: get_category_team_player
      produces:
      - application/json
      parameters:
        - name: category_id
          in: path
          required: true
          type: integer
        - name: team_id
          in: path
          required: true
          type: integer
        - name: player_id
          in: path
          required: true
          type: integer
          format: integer
      tags:
        - Player
        - Category
        - Team
      responses:
        200:
          description: pending
    post:
      summary: Create info of the summoned player of one team on one category
      description: Create info of the summoned player of one team on one category
      operationId: post_category_team_player
      tags:
        - Player
        - Category
        - Team
      produces:
      - application/json
      parameters:
      - name: category_id
        in: path
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        required: true
        type: integer
        format: integer
      - name: player_id
        in: path
        required: true
        type: integer
        format: integer
      - name: Body
        in: body
        required: false
        x-is-map: false
        schema:
          $ref: '#/definitions/CreateCategoryTeamPlayer'
      responses:
        200:
          description: Create info of the summoned player of one team on one category successful
      security: []
    put:
      summary: Update info of the summoned player of one team on one category
      description: Update info of the summoned player of one team on one category
      tags:
        - Player
        - Category
        - Team
      produces:
      - application/json
      parameters:
      - name: category_id
        in: path
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        required: true
        type: integer
        format: integer
      - name: player_id
        in: path
        required: true
        type: integer
        format: integer
      - name: Body
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/UpdateCategyTeamPlayer'
      responses:
        200:
          description: Update info of the summoned player of one team on one category successful
      security: []
  /category/{category_id}/team:
    get:
      summary: Gets all teams invited to participate on a category
      description: Gets all teams invited to participate on a category
      tags:
        - Category
        - Team
      produces:
      - application/json
      parameters:
        - name: category_id
          in: path
          description: Category ID
          required: true
          type: integer
          format: integer
      responses:
        200:
          description: results
      security: []
...
  /phase/{phase_id}/team:
    get:
      summary: Returns the teams participating in a phase
      description: Returns the teams participating in a phase
      tags:
        - Phase
      produces:
      - application/json
      parameters:
        - name: phase_id
          in: path
          description: Phase ID
          required: true
          type: integer
          format: integer
      responses:
        200:
          description: List of teams in a phase
      security: []
...
  /group/{group_id}/team:
    get:
      description: get teams invited to play on this group
      tags:
        - Group
        - Team
      produces:
      - application/json
      parameters:
        - name: group_id
          in: path
          description: Group ID
          required: true
          type: integer
          format: integer
      responses:
        200:
          description: list of teams currently on this group
      security: []
...
  /match/{match_id}/team/{team_id}/player:
    get:
      summary: Gets the player list for this team & match
      description: Gets the player list for this team & match
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: List of players assigned to this match & team
    put:
      summary: Toggles the participation of the given player in this match
      description: Toggles the participation of the given player in this match; necessary for updating the summoned player list for each match
      tags:
        - Match
        - Player
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      - name: player_data
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/MatchPlayerTeamRequest'
      responses:
        200:
          description: Correct update of match player list
  /match/{match_id}/team:
    get:
      summary: Gets all the info associated with the teams with the match. Summoned players included
      description: Gets all the info associated with the teams with the match. Summoned players included
      tags:
        - Match
        - Team
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      responses:
        200:
          description: (documentation pending)
...
  /match/:match_id/team/:team_id/player:
    post:
      summary: registers the starter eleven players for a match for each team
      description: registers the starter eleven players for a match for each team
      tags:
        - Match
      produces:
      - application/json
      parameters:
      - name: match_id
        in: path
        description: Match ID
        required: true
        type: integer
        format: integer
      - name: team_id
        in: path
        description: Team ID
        required: true
        type: integer
        format: integer
      - name: events
        in: body
        required: true
        schema:
          $ref: '#/definitions/initialPlayers'
      responses:
        200:
          description: (documentation pending)
...
  /team:
    get:
      summary: Gets the info of all teams
      description: Gets the info of all teams
      tags:
        - Team
      parameters:
        - name: category_type_id
          in: query
          description: Category Type ID
          required: false
          type: integer
          format: integer
        - name: gender_id
          in: query
          description: Gender ID
          required: false
          type: integer
          format: integer
        - name: subdiscipline_id
          in: query
          description: Subdiscipline ID
          required: false
          type: integer
          format: integer
      responses:
        200:
          description: List of teams
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a team
      description: Creates a team
      tags:
        - Team
      produces:
      - application/json
      parameters:
      - name: Authorization-Token
        in: header
        type: string
        required: true
      - name: Body
        in: body
        required: true
        x-is-map: false
        schema:
          $ref: '#/definitions/CreateTeamRequest'
      responses:
        200:
          description: team creation OK
      security: []
  /team/{team_id}:
    get:
      summary: Returns the detail of one team
      description: Returns the detail of one team
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
      tags:
        - Team
      responses:
        200:
          description: An array of teams
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Updates a team
      description: Updates a team
      parameters:
        - name: Authorization-Token
          in: header
          type: string
          required: true
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
        - name: team data
          in: body
          required: true
          x-is-map: false
          schema:
            $ref: '#/definitions/CreateTeamRequest'
      tags:
        - Team
      responses:
        200:
          description: Resource updated
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Logical delete of a team given the id
      description: Logical delete of a team given the id
      tags:
        - Team
      responses:
        200:
          description: object deleted
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
  /team/{team_id}/competition:
    get:
      summary: Returns the competitions where the given team is registered
      description: Returns the competitions where the given team is registered
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
      tags:
        - Team
        - Competition
      responses:
        200:
          description: An array of competitions
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /team/{team_id}/category/{category_id}/request:
    post:
      summary: Create one request for participated in the category from the team to participated
      description: Create one request for participated in the category from the team to participated
      operationId: Team_request_Category
      tags:
        - Team
      produces:
      - application/json
      parameters:
      - name: team_id
        in: path
        required: true
        type: integer
        format: integer
      - name: category_id
        in: path
        required: true
        type: integer
        format: integer
      responses:
        200:
          description:
      security: []
      x-unitTests:
      - request:
          method: POST
          uri: /team/30/category/5/request
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: NONE
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers: {}
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: Team_request_Category
        x-testDescription:
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
  /team/query/by_user:
    get:
      summary: Returns a list of teams on which this user is an owner
      description: Returns a list of teams on which this user is an owner
      tags:
        - Team
      parameters:
        - name: Authorization-Token
          in: header
          type: string
          required: true
      responses:
        200:
          description: List of teams on which this user is an owner
        403:
          description: Invalid Authorization-Token or no Authorization-Token received
        404:
          description: No teams found
  /team/{team_id}/match:
    get:
      summary: Returns all the matches for the team
      description: Returns all the matches for the team
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
      tags:
        - Team
      responses:
        200:
          description: An array of past, future matches and the next match for the team
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /team/{team_id}/event_calendar:
    get:
      summary: Returns the calendar events of one team
      description: Returns the calendar events of one team
      parameters:
        - name: team_id
          in: path
          description: Team ID
          required: true
          type: integer
          format: integer
      tags:
        - Team
      responses:
        200:
          description: the team information with an array of calendar events
        403:
          description: Invalid Authorization-Token or no Authorization-Token received
        404:
          description: No calendar events found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  MatchPlayerTeamRequest:
    title: Match Creation Request
    type: object
    properties:
      position:
        type: string
      number:
        type: integer
      player_id:
        type: integer
    required:
    - player_id
...
  UpdateCategyTeamPlayer:
    title: UpdateCategyTeamPlayer
    type: object
    properties:
      id:
        type: integer
        format: int32
      active:
        type: boolean
      number:
        type: integer
        format: int32
      position:
        type: string
    required:
    - id
  CreateCategoryTeamPlayer:
    title: CreateCategoryTeamPlayer
    type: object
    properties:
      category_id:
        type: string
      team_id:
        type: string
      player_id:
        type: string
      active:
        type: boolean
    required:
    - category_id
    - team_id
    - player_id
    - active
    - number
    - position
...
  initialPlayers:
      title: initialPlayers
      type: array
      items:
        type: object
        properties:
          number:
            type: integer
          player_id:
            type: integer
          match_id:
            type: integer
          team_id:
            type: integer
  Error:
    title: error
    type: object
    properties:
      description:
        type: string
...